home.admin/tor.requirements.txt
mechanize
Pillow
pytesseract
requests
PySocks
urwid